% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{getBarcode}
\alias{getBarcode}
\title{Read cell barcodes}
\usage{
getBarcode(barcodePath, verbose = TRUE)
}
\arguments{
\item{barcodePath}{Character(1). Path to a barcode file, such as a \code{.txt} file.}

\item{verbose}{Logical(1). Whether to print progress messages.}
}
\value{
A character vector of cell barcodes.
}
\description{
Reads a cell barcode file and returns the barcodes as a character vector.
}
\examples{
\dontrun{
  barcode <- getBarcode(
    barcodePath = "/path/to/barcode"
  )
  class(barcode)  # "character"
}

}
