% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGenome.R
\name{setGenomeAnnotation}
\alias{setGenomeAnnotation}
\title{Retrieve genome annotation}
\usage{
setGenomeAnnotation(object = NULL)
}
\arguments{
\item{object}{A \code{whereAreECDNAs} object, or \code{NULL}.}
}
\value{
The genome annotation object.
}
\description{
If an object is provided, returns its \verb{@genomeAnnotation} slot. If \code{object}
is \code{NULL}, retrieves the default annotation configured by
\code{\link[=addwhereAreGenomes]{addwhereAreGenomes()}}. An error is raised when no default genome is set.
}
\examples{
\dontrun{
  ## From an object
  annotation <- setGenomeAnnotation(object)

  ## From the globally configured default genome
  annotation <- setGenomeAnnotation()
}

}
